Taraban & McClelland (1987), Frequency by Consistency word set.

This data set gathered from appendix A includes 96 monosyllabic English words (48 word pairs). Each pair includes alternate pronunciations for provided inconsistent words. Within each pair words contrast in consistency (I vs. C), but are approximately matched for onset phoneme, word frequency, length and regularity.

tm1987a2

Format

A data.frame with 96 rows and 10 variables:

ortho

The orthographic word.

set

Item set. Maps word pairs, contrasting in consistency, but approximately matched for onset phoneme, word frequency, length and regularity.

altpron

Alternate pronunciations of inconsistent words.

wordfreq

Word Frequency.

samepron

Number of words with same orthographic word body having the same pronunciation.

MFA

Number of words where the same orthographic word body has the most frequent alternative pronunciation.

LFA

Number of words where the same orthographic word body has a less frequent alternative pronunciation.

frequency

Word frequency bin: H (high) or L (low).

regularity

Spelling regularity bin: R (regular)

consistency

Source

Taraban, R., & McClelland, J. L. (1987). Conspiracy effects in word pronunciation. Journal of Memory and Language, 26(6), 608-631. doi:Doi 10.1016/0749-596x(87)90105-7

Details

See source paper for details.

See also

tm1987a1